+++ context/cmdhandle/cmdflow.go
@@ -32,6 +32,20 @@ const (
 	ExitAlreadyRunning = 105
 )
 
+func SharedFolderExecuter(template configure.RunConfig, locationHandle func(string, string)) {
+	if len(template.Config.Use) > 0 {
+		GetLogger().WithField("uses", template.Config.Use).Info("shared executer")
+		for _, shared := range template.Config.Use {
+			externalPath := HandleUsecase(shared)
+			GetLogger().WithField("path", externalPath).Info("shared contxt location")
+			currentDir, _ := dirhandle.Current()
+			os.Chdir(externalPath)
+			locationHandle(externalPath, currentDir)
+			os.Chdir(currentDir)
+		}
+	}
+}
+
 func RunShared(targets string) {
 
 	allTargets := strings.Split(targets, ",")
@@ -71,7 +85,12 @@ func RunShared(targets string) {
 // RunTargets executes multiple targets
 func RunTargets(targets string, sharedRun bool) {
 
+	SetPH("CTX_TARGETS", targets)
+
 	if sharedRun {
+		// do it here makes sure we are not in the shared scope
+		currentDir, _ := dirhandle.Current()
+		SetPH("CTX_PWD", currentDir)
 		// run shared use
 		RunShared(targets)
 	}
@@ -122,6 +141,7 @@ func RunTargets(targets string, sharedRun bool) {
 	if !runSequencially {
 		// run in thread
 		for _, runTarget := range allTargets {
+			SetPH("CTX_TARGET", runTarget)
 			wg.Add(1)
 			fmt.Println(output.MessageCln(output.ForeBlue, "[exec:async] ", output.BoldTag, runTarget, " ", output.ForeWhite, templatePath))
 			go ExecuteTemplateWorker(&wg, true, runTarget, template)
@@ -130,6 +150,7 @@ func RunTargets(targets string, sharedRun bool) {
 	} else {
 		// trun one by one
 		for _, runTarget := range allTargets {
+			SetPH("CTX_TARGET", runTarget)
 			fmt.Println(output.MessageCln(output.ForeBlue, "[exec:seq] ", output.BoldTag, runTarget, " ", output.ForeWhite, templatePath))
 			exitCode := ExecPathFile(&wg, false, template, runTarget)
 			GetLogger().WithField("exitcode", exitCode).Info("RunTarget [Sequencially runmode] done with exitcode")
@@ -150,6 +171,15 @@ func setLogLevelByString(loglevel string) {
 }
 
 func checkRequirements(require configure.Require) (bool, string) {
+	// check operating system
+	if require.System != "" {
+		match := !StringMatchTest(require.System, configure.GetOs())
+		if !match {
+			return false, "operating system " + configure.GetOs() + " is not matching with " + require.System
+		}
+		return true, "matching os found " + require.System
+	}
+
 	// check file exists
 	for _, fileExists := range require.Exists {
 		fileExists = handlePlaceHolder(fileExists)
@@ -178,22 +208,63 @@ func checkRequirements(require configure.Require) (bool, string) {
 	}
 	// check environment variable is set
 
-	for name, value := range require.Environment {
-		envVar := os.Getenv(name)
-		if envVar != value {
-			return false, "environment variable[" + name + "] not matching with " + value
+	for name, pattern := range require.Environment {
+		envVar, envExists := os.LookupEnv(name)
+		if !envExists || !StringMatchTest(pattern, envVar) {
+			if envExists {
+				return false, "environment variable[" + name + "] not matching with " + pattern
+			}
+			return false, "environment variable[" + name + "] not exists"
 		}
 	}
 
 	// check variables
-	for name, value := range require.Variables {
-		defVar := GetPH(name)
-		if defVar != value {
-			return false, "runtime variable variable[" + name + "] not matching with " + value
+	for name, pattern := range require.Variables {
+		defVar, defExists := GetPHExists(name)
+		if !defExists || !StringMatchTest(pattern, defVar) {
+			if defExists {
+				return false, "runtime variable[" + name + "] not matching with " + pattern
+			}
+			return false, "runtime variable[" + name + "] not exists "
 		}
 	}
 
 	return true, ""
+}
+
+// StringMatchTest test a pattern and a value.
+// in this example: myvar: "=hello"
+// the patter is "=hello" and the value should be "hello" for a match
+func StringMatchTest(pattern, value string) bool {
+	first := pattern
+	maybeMatch := value
+	if len(pattern) > 1 {
+		maybeMatch = pattern[1:]
+		first = pattern[0:1]
+	}
+	switch first {
+	case "?":
+		GetLogger().WithFields(logrus.Fields{"pattern": pattern, "value": value, "check": first, "result": (value != "")}).Debug("check anything then empty")
+		return (value != "")
+	case "=":
+		GetLogger().WithFields(logrus.Fields{"pattern": maybeMatch, "value": value, "check": first, "result": (maybeMatch == value)}).Debug("check equal")
+		return (maybeMatch == value)
+	case "!":
+		GetLogger().WithFields(logrus.Fields{"pattern": maybeMatch, "value": value, "check": first, "result": (maybeMatch != value)}).Debug("check not equal")
+		return (maybeMatch != value)
+	case ">":
+		GetLogger().WithFields(logrus.Fields{"pattern": maybeMatch, "value": value, "check": first, "result": (maybeMatch < value)}).Debug("check greather then")
+		return (value > maybeMatch)
+	case "<":
+		GetLogger().WithFields(logrus.Fields{"pattern": maybeMatch, "value": value, "check": first, "result": (maybeMatch > value)}).Debug("check lower then")
+		return (value < maybeMatch)
+	case "*":
+		GetLogger().WithFields(logrus.Fields{"pattern": maybeMatch, "value": value, "check": first, "result": (maybeMatch > value)}).Debug("check lower then")
+		return (value != "")
+	default:
+		GetLogger().WithFields(logrus.Fields{"pattern": pattern, "value": value, "check": first, "result": (value != "")}).Debug("check equal against plain values")
+		return (pattern == value)
+	}
 }
 
 func listenerWatch(script configure.Task, target, logLine string, waitGroup *sync.WaitGroup, useWaitGroup bool, runCfg configure.RunConfig) {
@@ -302,6 +373,9 @@ func lineExecuter(waitGroup *sync.WaitGroup, useWaitGroup bool, stopReason confi
 		panelSize = script.Options.Panelsize
 	}
 	var mainCommand = defaultString(script.Options.Maincmd, DefaultCommandFallBack)
+	if configure.GetOs() == "windows" {
+		mainCommand = defaultString(script.Options.Maincmd, DefaultCommandFallBackWindows)
+	}
 	replacedLine := HandlePlaceHolder(codeLine)
 	if script.Options.Displaycmd {
 		fmt.Println(output.MessageCln(output.Dim, output.ForeYellow, " [cmd] ", output.ResetDim, output.ForeCyan, target, output.ForeDarkGrey, " \t :> ", output.BoldTag, output.ForeBlue, replacedLine))
@@ -450,6 +524,7 @@ func executeTemplate(waitGroup *sync.WaitGroup, useWaitGroup bool, runCfg config
 				if !canRun {
 					GetLogger().WithFields(logrus.Fields{
 						"target": target,
+						"reason": message,
 					}).Info("executeTemplate IGNORE because requirements not matching")
 					if script.Options.Displaycmd {
 						fmt.Println(output.MessageCln(output.ForeYellow, " [require] ", output.ForeBlue, message))